+++ packages/core/src/index.ts
@@ -3,15 +3,12 @@
  * @license     https://opensource.org/licenses/MIT
  */
 
-import { Interweave } from './Interweave';
-
 export * from './constants';
 export * from './Element';
 export * from './Filter';
+export * from './Interweave';
 export * from './Markup';
 export * from './match';
 export * from './Matcher';
 export * from './Parser';
 export * from './types';
-
-export default Interweave;