+++ packages/emoji/tests/Interweave.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Interweave, { Element, InterweaveProps } from 'interweave';
+import { Element, Interweave, InterweaveProps } from 'interweave';
 import { SOURCE_PROP } from 'interweave/test';
 import { render } from 'rut-dom';
 import { EmojiMatcher } from '../src/EmojiMatcher';